Make the Python html5lib tree-walker API use a dict with (namespace, name): value key: value pairs for attributes.
I promise I won't update this part of the API again — this is now something I'm happy with, so I won't do what I've just done and change it twice in six^W seven months again. The only possible slight update from this is to move to using an ordered dict for trees that can preserve attribute order, but this should have no API breakage.