Re-enable Genshi treewalker.
The Genshi treewalker was never fully updated when support for
namespaces was added to treewalkers. This finally fixes this, allowing
all tests to once again pass.

This removes the ignoring of children of void elements as the commonly
used treewalkers don't support this and there was some subtle bug I
couldn't see in the code. Given most don't support this, this seems
like little loss. (You will, however, still get an error from the
treewalker if a void element has children.)