Clean up html5lib module documentation
Right now the docs have entries for re-exports like
html5lib.__init__.HTMLParser, including full class documentation. This
is redundant with the docs for html5lib.html5parser.HTMLParser, which
is a public name anyway, so I think that it is best to be explicit that
this is a re-export.