Fix Travis CI builds on Python 3.4
Avoid passing -U to pip install commands. It causes the six package to
upgrade, which fails on the Travis 3.4 image. I cannot reproduce this
locally, so I'm left to conclude that the Travis Python 3.4 image is
corrupt in some way. The installed six package is perfectly fine for
testing, so use it without upgrading. All builds now pass.

Install six package first to avoid upgrading it then downgrading it in
some test configurations.