Collect PyPy and AppVeyor coverage using tox
Coverage of PyPy was disabled however it works
without trouble, albeit a little slow.

To allow tox to run tests under coverage,
and with arguments passed to coverage,
PYTEST_COMMAND can now be used to override
the default command used to run the tests.

Due to pips inability to deal with multiple
requirements for the same package, six==1.9 is
forcably installed after the tox environment
has been setup.