Remove undoc-members and add documentation for __init__
The public API should all be documented. Anything that isn't documented with a
docstring or starts with a _ isn't part of the public API. Given that, we
shouldn't be autodoc'ing undocumented members.

We do want to document __init__ since that's how the classes get built. So
we explicitly add that to autodoc.

I think this is a good base to build on. If it isn't, we can adjust things
and maybe explicitly specify what should and shouldn't be documented.