Fix Pytest4.x compatibility error
According to pytest docs:
```
marks in pytest.mark.parametrize
Removed in version 4.0.

Applying marks to values of a pytest.mark.parametrize call is now deprecated.
...

This was considered hard to read and understand, and also its implementation
presented problems to the code preventing further internal improvements in the
marks architecture.

To update the code, use pytest.param
```

Signed-off-by: Stanislav Levin <slev@altlinux.org>