Test more on GitHub Actions
The only environments left on AppVeyor are now 3.5 and 3.6, which aren't
supported on GitHub Actions.

The remaining Python versions are now also tested on Windows using GitHub
Actions.

Additionally, a new group of dependencies called "oldest" is now tested as well
to ensure compatibility with the oldest pinned versions in requirements.txt and
setup.py.